-update delete for broken files

+++ blackvue.py
@@ -155,7 +155,8 @@ async def downloadFilesToTmpFolder():
             videoFileTransferList.append(videoFile)                             
 
         except Exception as e:
-            os.remove(filePath)
+            if os.path.exists(filePath):
+                os.remove(filePath)
             print(e, flush=True)
 
 async def main():
